% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latex2Rmd}
\alias{latex2Rmd}
\title{Latex to R markdown converter}
\usage{
latex2Rmd(input, output)
}
\arguments{
\item{input}{Character Filename of input file}

\item{output}{character Filename of output file}
}
\description{
Does 70 percent of the job, the rest is manual edits.
}
\details{
I recommend to first transfer the content of the latex file to a RStudio Rmd template (removing headers, packages importts, etc) and then run the function.

What it doesn't do:
\itemize{
 \item Insert figures from external sources.
 \item Differentiate ordered lists from unordered lists
 \item Fix references to figures in the text
 \item Adapt file headers and so on
}
}
\author{
Loic Dutrieux
}

